Restructuration + DRY

+++ functions.php
@@ -5,8 +5,10 @@
     print_r($a);
     echo '</pre>';
   }
-
-  // Generate a random string 
+  
+  function check($val) : bool{
+    return isset($val) && !empty($val);
+  }
 
   function generateRandomString($length = 8) {
     $characters = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
@@ -23,7 +25,7 @@
     return($myfiles);
   }
   
-  function upload_files($filename, $path_to_img, $extensions_autorisees = ['jpg', 'jpeg', 'png']){
+  function upload_files($filename, $path_to_img, $extensions_autorisees = ['jpg', 'jpeg', 'png'], $replace = false){
 
     foreach($_FILES as $img_name => $img_prop)
     {
@@ -32,8 +34,10 @@
 
       if (image_has_valid_size($img_prop) && image_has_valid_ext($img_prop, $extensions_autorisees)){
         $imageFilename = getFiles($path_to_img.$filename.'');
-        unlink($path_to_img.$filename.'/'.$imageFilename[2].'');
-        unset($imageFilename['2']);
+        if($replace){
+          unlink($path_to_img.$filename.'/'.$imageFilename[2].'');
+        }
+        unset($imageFilename[2]);
         move_uploaded_file($img_prop['tmp_name'], $path_to_img.$filename.'/' . basename($img_prop['name'])); 
       }
       
@@ -42,7 +46,7 @@
 
   function image_has_valid_size($img_prop)
   {
-    return $img_prop['size'] <= 1000000000000000;
+    return $img_prop['size'] <= 1000000000;
   }
 
   function image_has_valid_ext($img_prop, $valid_ext) : bool 
@@ -50,6 +54,6 @@
     $infosfichier = pathinfo($img_prop['name']);
     $extension_upload = $infosfichier['extension'];
     return in_array($extension_upload, $valid_ext);
-    
+   
   }
 ?>
